use more cpus for awb encoding stage, make light model default output format, make split mode with 20mins default, be less talkative

+++ createPlayawayAudioBook.pl
@@ -1,6 +1,6 @@
 #!/usr/bin/perl
 use strict;
-my $version="1.05";
+my $version="1.06";
 
 # this script assumes everything is in the directory from which it was started:
 #  - the script itself
@@ -10,6 +10,7 @@ my $version="1.05";
 # ffmpeg, ffprobe and mktemp must be path available
 #
 # Version history
+# V1.06   : use more cpus for awb encoding stage, make light model default output format, make split mode with 20mins default, be less talkative
 # V1.05   : added auto bitrate selection based on flash storage of device and runtime of files, added file split mode
 # V1.04   : updated for faster none wine encoder, thanks to Dhiru Kohlia https://github.com/kholia/amr-wbp-plus
 # V1.03   : added option to choose to generate files for Playaway light or classic, cosmetics and more error checking
@@ -17,8 +18,16 @@ my $version="1.05";
 # V1.01   : changes in COP field and ffmpeg parameters
 # V1.0    : initial version
 
-print "Playaway audio book creator $version\n\n";
-my @flashSizes=("105000","229000","469000");#size of usable 1k blocks
+$|=1;
+my $cpus = do { local @ARGV='/proc/cpuinfo'; grep /^processor\s+:/, <>;};
+if($cpus !~ /^\d+$/){$cpus=1;}
+my $maxChilds=$cpus;
+eval{#be nice and encode in background
+    setpriority(0, 0, 20);
+};
+my $encoderCmd="./encoder-new";
+print "Playaway audio book creator $version (using $maxChilds CPUs) \n\n";
+my @flashSizes=("100000","227000","467000");#size of usable 1k blocks
 preFlightChecks();
 print "Reading file list";
 my @files=();
@@ -54,6 +63,8 @@ if($bitRate=~/^a\d$/){
     }
     if($bitRate<10){
         print STDERR "ERROR: The files do not fit completely on that flash.\n";
+        print STDERR "Try splitting the file into multiple parts and convert them individually.\n";
+        print STDERR "The following command can be used for lossless splitting (e.g. split every 10hrs):\n   ffmpeg -i input.amr -c copy -map 0 -segment_time 10:00:00 -f segment output\%03d.amr\n";
         exit 2;
     }
     print "Using automatic determined bitrate: $bitRate\n";
@@ -63,13 +74,13 @@ if($bitRate=~/^a\d$/){
         exit 2;
     }
 }
-print "Splitmode for files? ([n],5...60): ";my $splitMode=<>;chomp($splitMode);if($splitMode eq ""){$splitMode="n";};
+print "Splitmode for files? (n,5...60) [20]: ";my $splitMode=<>;chomp($splitMode);if($splitMode eq ""){$splitMode="20";};
 if($splitMode ne "n"){
     if(($splitMode!~/^\d+$/)||($splitMode < 5)||($splitMode > 60)){
          print STDERR "Only values between 5...60 are allowed.\n";
     }
 }
-print "Generate for Playway model classic or light? [c]/l: ";my $model=<>;chomp($model);if($model eq ""){$model="c";};
+print "Generate for Playway model classic or light? c/[l]: ";my $model=<>;chomp($model);if($model eq ""){$model="l";};
 if(($model ne "c")&&($model ne "l")){
     print STDERR "ERROR: Model selection must eihter be c or l.\n";
     exit 2;
@@ -87,6 +98,7 @@ if(-d $bookName){
 
 my $origFileCnt=0;
 my @convertedFileRuntimes=();
+my $childCnt=0;
 foreach(@files){
     my $origFile=$_;
     $origFileCnt++;
@@ -99,11 +111,11 @@ foreach(@files){
         if($origFile!~/\.wav$/){
             $tempFile=`mktemp -u -p "."`;chomp($tempFile);$tempFile.".wav";
             if($splitMode eq "n"){
-                $back=system("ffmpeg -hide_banner -v error -stats -i \"$origFile\" -f wav -c:a pcm_s16le -ar 44100 -ac 1 -empty_hdlr_name 1 -fflags +bitexact -flags:v +bitexact -flags:a +bitexact -map_metadata -1 \"$tempFile\"");
+                $back=system("ffmpeg -hide_banner -v error -stats -i \"$origFile\" -f wav -c:a pcm_s16le -ar 44100 -ac 1 -empty_hdlr_name 1 -fflags +bitexact -flags:v +bitexact -flags:a +bitexact -map_metadata -1 \"$tempFile\" >> /dev/null 2>&1");
                 $endFlag=1;
             }else{
                 print "Current timecode ".$currentPartStart."min\n";
-                $back=system("ffmpeg -hide_banner -v error -stats -ss ".($currentPartStart*60)." -t ".($splitMode*60)." -i \"$origFile\" -f wav -c:a pcm_s16le -ar 44100 -ac 1 -empty_hdlr_name 1 -fflags +bitexact -flags:v +bitexact -flags:a +bitexact -map_metadata -1 \"$tempFile\"");
+                $back=system("ffmpeg -hide_banner -v error -stats -ss ".($currentPartStart*60)." -t ".($splitMode*60)." -i \"$origFile\" -f wav -c:a pcm_s16le -ar 44100 -ac 1 -empty_hdlr_name 1 -fflags +bitexact -flags:v +bitexact -flags:a +bitexact -map_metadata -1 \"$tempFile\" >> /dev/null 2>&1");
                 $currentPartStart+=$splitMode;
             }
         }else{
@@ -119,12 +131,22 @@ foreach(@files){
             if($fileRuntime>0){
                 push @convertedFileRuntimes,$fileRuntime;
                 my $bookFile="$bookName/".sprintf("%04d", scalar(@convertedFileRuntimes))." $bookName 0000.awb";
-                $back=system("./encoder-new -rate $bitRate -mono -ff raw -if \"$tempFile\" -of \"$bookFile\"");
-                if($origFile ne $tempFile){
-                    unlink $tempFile;
-                }
-                if((stat($bookFile))[7] == 0){#encoder-new does not ouput useful return codes so check the file size
-                    print STDERR "ERROR: Could not convert temporary wav file to awb plus Playaway file.\n";
+                my $pid=fork();
+                if($pid){#parent
+                    $childCnt++;
+                    if($childCnt>=$maxChilds){
+                        select(undef,undef,undef,0.1);
+                        waitpid -1,0;#wait for one child to finish
+                        $childCnt--;
+                    }
+                }elsif($pid==0){#child
+                    $back=system("$encoderCmd -rate $bitRate -mono -ff raw -if \"$tempFile\" -of \"$bookFile\" >> /dev/null 2>&1");
+                    if($origFile ne $tempFile){
+                        unlink $tempFile;
+                    }
+                    exit 0;#child just quits after finishing encoding
+                }else{#error
+                    print STDERR "Could not fork.\n";
                     exit 7;
                 }
             }else{
@@ -138,6 +160,28 @@ foreach(@files){
         }
     }while($endFlag==0);
 }
+while($childCnt>0){#wait for all childs to finish
+    select(undef,undef,undef,0.1);
+    waitpid -1,0;#wait for one child to finish
+    $childCnt--;
+}
+
+#verify if all files could be converted
+if(opendir(D,$bookName)){
+    my @f=readdir(D);
+    foreach(@f){
+        if(($_ ne ".")&&($_ ne "..")){
+            if((stat($bookName."/".$_))[7] == 0){#encoder-new does not ouput useful return codes so check the file size
+                print STDERR "ERROR: Could not convert file to awb plus Playaway file.\n";
+                exit 8;
+            }
+        }
+    }
+}else{
+    print STDERR "Error checking converted files after converting. Could not open directory $bookName.\n";
+    exit 8;
+}
+
 my $preAllModels="AWBVOL082002SLD090080PUP003NMD";
 my $preLight=$preAllModels.sprintf("%03d",scalar(@convertedFileRuntimes))."BLN020BLP020ELA";
 my $preClassic=$preAllModels.sprintf("%03d",scalar(@convertedFileRuntimes));
@@ -160,9 +204,9 @@ if(open(P,">$bookName/PATWEAKS.DAT")){
     close(P);
 }else{
     print STDERR "ERROR: Could not create $bookName/PATWEAKS.DAT.\n";
-    exit 8;
+    exit 9;
 }
-
+print "Finished. No errors detected.\n";
 exit;
 
 sub findFiles{
@@ -203,9 +247,13 @@ sub preFlightChecks{
         print STDERR "ERROR: mktemp does not seem to work.\n";
         exit 20;
     }
-    $back=system("./encoder-new > /dev/null 2>&1");
+    $back=system("$encoderCmd > /dev/null 2>&1");
     if($back != 256){#exit code 1, which is normal exit code when no encoding happens
-        print STDERR "ERROR: encoder-new binary not found.\n";
-        exit 20;
+        $encoderCmd="encoder-new";#try to look in path for encoder
+        $back=system("$encoderCmd > /dev/null 2>&1");
+        if($back != 256){#exit code 1, which is normal exit code when no encoding happens
+            print STDERR "ERROR: encoder-new binary not found.\n";
+            exit 20;
+        }
     }
 }
